+++ config/environment.rb
@@ -1,8 +1,9 @@
 require 'pry'
 
-# require_relative '../lib/concerns/findable'
-# require_relative '../lib/concerns/memorable'
-# require_relative '../lib/concerns/paramable'
+require_relative '../lib/concerns/findable.rb'
+require_relative '../lib/concerns/memorable.rb'
+require_relative '../lib/concerns/paramable.rb'
+# require_relative '../lib/concerns/initialize.rb'
 
 require_relative '../lib/artist.rb'
 require_relative '../lib/song.rb'
