+++ lib/concerns/findable.rb
@@ -1,0 +1,9 @@
+require "pry"
+
+module Findable
+
+  def find_by_name(name)
+    self.all.detect {|i| i.name == name}
+  end
+
+end
