+++ lib/concerns/initialize.rb
@@ -1,0 +1,9 @@
+
+
+module Initialize
+
+  def initialize
+    self.class.all << all
+  end
+  
+end
