+++ lib/concerns/paramable.rb
@@ -1,0 +1,10 @@
+require "pry"
+
+module PARAMABLE
+
+  def to_param
+    name.downcase.gsub(' ', '-')
+    # binding.pry
+  end
+
+end
