+++ lib/song.rb
@@ -1,36 +1,18 @@
 require 'pry'
 
 class Song
-  attr_accessor :name
-  attr_reader :artist
+
+  # extend Memorable
+  extend Memorable::ClassMethods
+  include Memorable::InstanceMethods
+  extend Findable
+  include PARAMABLE
+  attr_accessor :name, :artist
 
   @@songs = []
-
-  def initialize
-    @@songs << self
-  end
-
-  def self.find_by_name(name)
-    @@songs.detect{|a| a.name == name}
-  end
-
-  def artist=(artist)
-    @artist = artist
-  end
-
-  def self.reset_all
-    @@songs.clear
-  end
-
-  def to_param
-    name.downcase.gsub(' ', '-')
-  end
-
-  def self.count
-    self.all.count
-  end
 
   def self.all
     @@songs
   end
+  
 end
